// src/generated/models/Ppa_intakelogsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ppa_intakelogsppa_injectionsite = {
  894250000: 'RightHip',
  894250001: 'LeftHip',
  894250002: 'AbdominalRight',
  894250003: 'AbdominalCenter',
  894250004: 'AbdominalLeft'
} as const;
export type Ppa_intakelogsppa_injectionsite = keyof typeof Ppa_intakelogsppa_injectionsite;
export const Ppa_intakelogsppa_status = {
  894250000: 'Taken',
  894250001: 'Skipped',
  894250002: 'Missed'
} as const;
export type Ppa_intakelogsppa_status = keyof typeof Ppa_intakelogsppa_status;
export const Ppa_intakelogsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ppa_intakelogsstatecode = keyof typeof Ppa_intakelogsstatecode;
export const Ppa_intakelogsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ppa_intakelogsstatuscode = keyof typeof Ppa_intakelogsstatuscode;

export interface Ppa_intakelogsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  ppa_injectionsite?: Ppa_intakelogsppa_injectionsite;
  ppa_intakelogid: string;
  ppa_loggedat: string;
  ppa_logname: string;
  "ppa_Medication@odata.bind": string;
  ppa_notes?: string;
  ppa_photourl?: string;
  ppa_scheduledfor: string;
  ppa_status: Ppa_intakelogsppa_status;
  statecode: Ppa_intakelogsstatecode;
  statuscode?: Ppa_intakelogsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ppa_intakelogs extends Ppa_intakelogsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  ppa_injectionsitename?: string;
  ppa_medicationname?: string;
  ppa_statusname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  ppa_medication?: object;
  _ppa_medication_value?: string;
}